pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-manager"
version = "0.1.0"
description = "Project Structure Manager for generating standardized business project folders"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.5",
    "pyinstaller>=6.16.0",
    "python-docx>=1.2.0",
    "sqlalchemy>=2.0.0",
]

[project.scripts]
project-manager = "__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.pyinstaller]
main = "__main__.py"
